+++ desapp-groupc1h-backend/src/test/java/com/viandasya/model/user/ServiceProfileTest.java
@@ -2,6 +2,7 @@ package com.viandasya.model.user;
 
 import com.viandasya.model.menu.Menu;
 
+import com.viandasya.model.timeslot.TimeSlot;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -11,6 +12,7 @@ import org.mockito.junit.MockitoJUnitRunner;
 import java.util.ArrayList;
 import java.util.List;
 
+import static com.viandasya.model.builders.DateTimeSlotBuilder.anyDateTimeSlot;
 import static com.viandasya.model.builders.ServiceProfileBuilder.anyServiceProfile;
 
 @RunWith(MockitoJUnitRunner.class)
@@ -18,14 +20,23 @@ public class ServiceProfileTest {
 
     @Test
     public void testHas20ValidMenus_serviceWithValid20MenusGivesTrue() {
-        ServiceProfile service = anyServiceProfile().setMenus(this.createNMockValidMenus(20)).createServiceProfile();
+        ServiceProfile serviceProfile = anyServiceProfile().createServiceProfile();
+        TimeSlot timeSlot = anyDateTimeSlot().createDateTimeSlot();
+        List<Menu> menus = this.createNMockValidMenus(20);
+        Balance balance = new Balance(0);
+        Service service = new Service(serviceProfile, timeSlot, menus,balance);
+
 
         Assert.assertTrue(service.has20ValidMenus());
     }
 
     @Test
     public void testHas20ValidMenus_serviceWith17ValidMenusGivesFalse() {
-        ServiceProfile service = anyServiceProfile().setMenus(this.createNMockValidMenus(17)).createServiceProfile();
+        ServiceProfile serviceProfile = anyServiceProfile().createServiceProfile();
+        TimeSlot timeSlot = anyDateTimeSlot().createDateTimeSlot();
+        List<Menu> menus = this.createNMockValidMenus(17);
+        Balance balance = new Balance(0);
+        Service service = new Service(serviceProfile, timeSlot, menus,balance);
 
         Assert.assertFalse(service.has20ValidMenus());
     }
